move email scan stuff into base schema file
