+++ api-js/src/types/base/scan/index.js
@@ -1,8 +1,6 @@
-const { emailScanType } = require('./email-scan')
 const { webScanType, webScanConnection } = require('./web-scan')
 
 module.exports = {
-  emailScanType,
   webScanType,
   webScanConnection,
 }
